+++ src/Command.h
@@ -11,7 +11,7 @@ public:
   // Compare the name of the command (argument #0) to the specified name
   bool is(const char *name) { return this->_arguments.size() > 0 && strcmp(this->_arguments.at(0).c_str(), name) == 0; }
   // Get the number of arguments EXCLUDING THE COMMAND NAME
-  int arguments() { return this->_arguments.size() - 1; }
+  size_t arguments() { return this->_arguments.size() - 1; }
   // Get the nth argument. If this more than the actual number of arguments. return the default value specified
-  String arg(int n, const char *defaultValue);
+  String arg(size_t n, const char *defaultValue);
 };
